Wave 6: CI quality gates — golangci-lint, gofmt, frontend lint+test
- .golangci.yml (v2): errcheck/govet/staticcheck/unused/ineffassign +
  misspell/unconvert/bodyclose; narrow value-less-error exclusions only.
  Fixing findings caught a real masked bug (mail body never reached its
  event) and hardened TSDB durability error handling.
- CI: new 'lint' job (gofmt check + golangci-lint); 'ui' job now also runs
  frontend lint + vitest before building the embedded dist. Combined with
  the existing race-tests, PG matrix, cross-build and types drift gate, the
  pipeline now guards Go + TS formatting, linting, types, and tests on every PR.

golangci-lint: 0 issues. gofmt: clean. go test + npm test: green.

+++ cmd/np-agent/sys_unix.go
@@ -57,8 +57,10 @@ func diskUsage(mount string) (float64, uint64, bool) {
 	if err := syscall.Statfs(mount, &st); err != nil {
 		return 0, 0, false
 	}
-	total := uint64(st.Blocks) * uint64(st.Bsize)
-	free := uint64(st.Bavail) * uint64(st.Bsize)
+	// st.Blocks/st.Bavail are uint64 on both darwin and linux; st.Bsize is
+	// uint32 (darwin) / int64 (linux), so it still needs a conversion.
+	total := st.Blocks * uint64(st.Bsize)
+	free := st.Bavail * uint64(st.Bsize)
 	if total == 0 {
 		return 0, 0, false
 	}
